feat(Tree): preselected items and export tree helpers
