+++ src/components/Tree/TreeItem/index.ts
@@ -5,3 +5,4 @@ export * from './ExpandButton';
 export * from './TreeItem';
 export * from './TreeItemOverlay';
 export * from './useTreeItem';
+export * from './TreeItemMultiselect';
