+++ src/components/Tree/helpers/index.ts
@@ -7,3 +7,4 @@ export * from './reducer';
 export * from './constants';
 export * from './treeHandleKeyDown';
 export * from './sensorsActivationConstraint';
+export * from './multiselect';
